llm_inventory_agent/requirements.txt
paho-mqtt==1.6.1
requests==2.32.3
llama-cpp-python==0.2.83
